Add jutsu prefill to combat simulator
