+++ admin/combat_simulator/vs.php
@@ -6,9 +6,6 @@ require __DIR__ . "/../_authenticate_admin.php";
 require __DIR__ . "/TestFighter.php";
 require __DIR__ . "/calcDamage.php";
 
-/** @var string[] $bloodline_combat_boosts */
-require_once __DIR__ . '/../constraints/bloodline.php';
-
 $rankManager = new RankManager($system);
 $rankManager->loadRanks();
 
@@ -82,42 +79,15 @@ if(isset($_POST['run_simulation'])) {
     }
 }
 
-$bloodlines_by_rank = [];
-$result = $system->db->query("SELECT * FROM `bloodlines` WHERE `rank` < 5 ORDER BY `rank` ASC");
-if ($system->db->last_num_rows > 0) {
-    while ($row = $system->db->fetch($result)) {
-        $bloodlines_by_rank[$row['rank']][$row['bloodline_id']] = Bloodline::fromArray($row);
-    }
-}
+require 'vs_fighter_input.php';
 
 ?>
 
 <style>
-    label {
-        display: inline-block;
-    }
-
     .vs_container {
         width: 800px;
         margin: 5px auto;
         text-align: center;
-    }
-
-    .versusFighterInput {
-        width:300px;
-        display:inline-block;
-        border:1px solid #000000;
-        border-radius:10px;
-
-        text-align: left;
-
-        padding: 8px;
-
-        background-color: #f2f2f2;
-    }
-
-    .jutsu_input {
-        margin: 5px auto;
     }
 
     input[type='submit'] {
@@ -126,153 +96,10 @@ if ($system->db->last_num_rows > 0) {
     }
 </style>
 
-<script>
-    function prefillBloodline(fighterKey) {
-        let bloodlineSelect = document.getElementById(`bloodline_prefill_${fighterKey}`);
-        if(bloodlineSelect == null) {
-            console.error("Invalid bloodline select element");
-            return;
-        }
-
-        let selectedBloodline = bloodlineSelect.selectedOptions[0];
-        if(parseInt(selectedBloodline.value) === 0) {
-            console.warn("No bloodline selected!");
-            return;
-        }
-
-        let selectedBloodlineBoosts = JSON.parse(
-            bloodlineSelect.selectedOptions[0].getAttribute("data-boosts")
-        );
-
-        selectedBloodlineBoosts.forEach((boost, i) => {
-            let boostEl = document.getElementById(`${fighterKey}_bloodline_boost_${i + 1}`);
-            let boostAmountEl = document.getElementById(`${fighterKey}_bloodline_boost_${i + 1}_amount`);
-
-            console.log(`${fighterKey}_bloodline_boost_${i + 1}`, boostEl);
-            console.log(`${fighterKey}_bloodline_boost_${i + 1}_amount`, boostAmountEl);
-
-            boostEl.value = boost.effect;
-            boostAmountEl.value = boost.power;
-        });
-    }
-</script>
-
-<?php
-    /**
-     * @param string $fighter_form_key
-     * @param array  $bloodline_combat_boosts
-     * @param Bloodline[][]  $bloodlines_by_rank
-     * @return void
-     */
-    function displayFighterInput(string $fighter_form_key, array $bloodline_combat_boosts, array $bloodlines_by_rank): void {
-        $stats = [
-            'ninjutsu_skill',
-            'taijutsu_skill',
-            'genjutsu_skill',
-            'bloodline_skill',
-            'speed',
-            'cast_speed',
-        ];
-        ?>
-        <div class='versusFighterInput' style='margin-left: 20px;'>
-            <b><?= ucwords($fighter_form_key) ?></b><br />
-            <?php foreach($stats as $stat): ?>
-                <label style='width:110px;'><?= $stat ?>:</label>
-                <input type='number' step='10' name='<?= $fighter_form_key ?>[<?= $stat ?>]' value='<?= $_POST[$fighter_form_key][$stat] ?? 0 ?>' /><br />
-            <?php endforeach; ?>
-            <br />
-
-            Bloodline boosts<br />
-            <select
-                id='bloodline_prefill_<?= $fighter_form_key ?>'
-                name='<?= $fighter_form_key ?>_bloodline_id'
-                onchange='prefillBloodline("<?= $fighter_form_key ?>")'
-            >
-                <option value='0'>Select to auto-fill boosts</option>
-                <?php foreach($bloodlines_by_rank as $rank => $bloodlines): ?>
-                    <optgroup label="<?= Bloodline::$public_ranks[$rank] ?>">
-                        <?php foreach($bloodlines as $bloodline): ?>
-                            <option
-                                value='<?= $bloodline->bloodline_id ?>'
-                                data-boosts='<?= json_encode($bloodline->base_combat_boosts) ?>'
-                                <?= (
-                                    ($_POST["{$fighter_form_key}_bloodline_id"] ?? '') == $bloodline->bloodline_id
-                                        ? "selected='selected'"
-                                        : ''
-                                ) ?>
-                            >
-                                <?= $bloodline->name ?>
-                            </option>
-                        <?php endforeach; ?>
-                    </optgroup>
-                <?php endforeach; ?>
-            </select><br />
-            <br />
-            <?php for($i = 1; $i <= 3; $i++): ?>
-                <select
-                    id='<?= $fighter_form_key?>_bloodline_boost_<?= $i ?>'
-                    name='<?= $fighter_form_key ?>[bloodline_boost_<?= $i ?>]'
-                >
-                    <option value='none'>None</option>
-                    <?php foreach($bloodline_combat_boosts as $boost): ?>
-                        <option value='<?= $boost ?>'
-                            <?= (($_POST[$fighter_form_key]["bloodline_boost_{$i}"] ?? '') == $boost ? "selected='selected'" : '') ?>
-                        >
-                            <?= $boost ?>
-                        </option>
-                    <?php endforeach; ?>
-                </select>
-                <input
-                    type='number'
-                    id='<?= $fighter_form_key?>_bloodline_boost_<?= $i ?>_amount'
-                    name='<?= $fighter_form_key ?>[bloodline_boost_<?= $i ?>_power]'
-                    style='width:60px'
-                    value='<?= $_POST[$fighter_form_key]["bloodline_boost_{$i}_power"] ?? 0 ?>'
-                />
-                <br />
-            <?php endfor; ?>
-            <br />
-            <br />
-
-            <div class='jutsu_input'>
-                <label style='width:110px;'>Jutsu power:</label>
-                <input type='text' name='<?= $fighter_form_key ?>[jutsu_power]' value='<?= $_POST[$fighter_form_key]['jutsu_power'] ?? 1 ?>' /><br />
-                <label>
-                    <input type='radio' name='<?= $fighter_form_key ?>[jutsu_type]' value='ninjutsu'
-                        <?= (($_POST[$fighter_form_key]['jutsu_type'] ?? '') == 'ninjutsu' ? "checked='checked'" : '') ?>
-                    />
-                    Ninjutsu
-                </label><br />
-                <label>
-                    <input type='radio' name='<?= $fighter_form_key ?>[jutsu_type]' value='taijutsu'
-                        <?= (($_POST[$fighter_form_key]['jutsu_type'] ?? '') == 'taijutsu' ? "checked='checked'" : '') ?>
-                    />
-                    Taijutsu
-                </label><br />
-                <label>
-                    <input type='radio' name='<?= $fighter_form_key ?>[jutsu_type]' value='genjutsu'
-                        <?= (($_POST[$fighter_form_key]['jutsu_type'] ?? '') == 'genjutsu' ? "checked='checked'" : '') ?>
-                    />
-                    Genjutsu
-                </label><br />
-            </div>
-        </div>
-        <?php
-    }
-?>
-
 <form action='vs.php' method='post'>
     <div class='vs_container'>
-        <?php displayFighterInput(
-            fighter_form_key: 'fighter1',
-            bloodline_combat_boosts: $bloodline_combat_boosts,
-            bloodlines_by_rank: $bloodlines_by_rank
-        ); ?>
-        <?php displayFighterInput(
-            fighter_form_key: 'fighter2',
-            bloodline_combat_boosts: $bloodline_combat_boosts,
-            bloodlines_by_rank: $bloodlines_by_rank
-        ); ?>
+        <?php displayFighterInput(system: $system, fighter_form_key: 'fighter1'); ?>
+        <?php displayFighterInput(system: $system, fighter_form_key: 'fighter2'); ?>
         <input type='submit' name='run_simulation' value='Run Simulation' />
     </div>
 </form>
